[#4441] Added alt statements and UTs
MySql and PostgreSQL lease mgrs now call
alteranate SQL statements when lease is
under SFLQ jurisdiction

/src/hooks/dhcp/mysql/mysql_lease_mgr.*
    Added alternate statements:
    {MySqlLeaseMgr::SFLQ_INSERT_LEASE4,
    {MySqlLeaseMgr::SFLQ_UPDATE_LEASE4,
    {MySqlLeaseMgr::SFLQ_DELETE_LEASE4,
    {MySqlLeaseMgr::SFLQ_INSERT_LEASE6,
    {MySqlLeaseMgr::SFLQ_UPDATE_LEASE6,
    {MySqlLeaseMgr::SFLQ_DELETE_LEASE6,

    Modified to call alt statements:
    MySqlLeaseMgr::addLease*()
    MySqlLeaseMgr::udpateLease*()
    MySqlLeaseMgr::deleteLease*()

    MySqlLeaseMgr::sflqPickFreeLease4() fix output bind

/src/hooks/dhcp/mysql/tests/mysql_lease_mgr_unittest.cc
    Renamed testSFLQ4() to testSflqCreateAndPick4()
    Renamed testSFLQ6() to testSflqCreateAndPick6()

    TEST_F(MySqlLeaseMgrTest, testSflqLeaseOps4 - new test
    TEST_F(MySqlLeaseMgrTest, testSflqLeaseOps6 - new test

/src/hooks/dhcp/pgsql/pgsql_lease_mgr.*
    Added alternate statements:
        SFLQ_INSERT_LEASE4
        SFLQ_UPDATE_LEASE4
        SLQ_DELETE_LEASE4
        SFLQ_INSERT_LEASE6
        SFLQ_UPDATE_LEASE6
        SFLQ_DELETE_LEASE6

    Modified to call alt statements:
    PgSqlLeaseMgr::addLease*()
    PgSqlLeaseMgr::udpateLease*()
    PgSqlLeaseMgr::deleteLease*()

/src/hooks/dhcp/pgsql/tests/pgsql_lease_mgr_unittest.cc
    Renamed testSFLQ4() to testSflqCreateAndPick4()
    Renamed testSFLQ6() to testSflqCreateAndPick6()

    TEST_F(PgSqlLeaseMgrTest, testSflqLeaseOps4 - new test
    TEST_F(PgSqlLeaseMgrTest, testSflqLeaseOps6 - new test

/src/lib/dhcpsrv/lease_mgr.*
    bool LeaseMgr::useSharedFlqStatement(*) - new functions

/src/lib/dhcpsrv/testutils/generic_lease_mgr_unittest.*
    Renamed testSFLQ4() to testSflqCreateAndPick4()
    Renamed testSFLQ6() to testSflqCreateAndPick6()

    testSflqLeaseOps4() - new test
    testSflqLeaseOps6()- new test

/src/share/database/scripts/pgsql/dhcpdb_create.pgsql
/src/share/database/scripts/pgsql/upgrade_032_to_033.sh.in
    Fixed null returns in sflqPickFreeLease*

+++ src/hooks/dhcp/pgsql/tests/pgsql_lease_mgr_unittest.cc
@@ -81,8 +81,15 @@ public:
     /// @brief Destroys the LM and the schema.
     void destroyTest() {
         LeaseMgrFactory::destroy();
+
         // If data wipe enabled, delete transient data otherwise destroy the schema
-        destroyPgSQLSchema();
+        if (getenv("KEA_UNIT_TEST_KEEP_DB_DATA")) {
+            // Leaves schema intact for post-test debugging.
+            std::cout << "KEA_UNIT_TEST_KEEP_DB_DATA set" << std::endl;
+        } else {
+            // If data wipe enabled, delete transient data otherwise destroy the schema
+            destroyPgSQLSchema();
+        }
 
         // Disable Multi-Threading.
         MultiThreadingMgr::instance().setMode(false);
@@ -1399,12 +1406,20 @@ TEST_F(PgSqlLeaseMgrTest, updateStatsOn6DifferentSubnetPD) {
     testUpdateStatsOn6DifferentSubnetPD();
 }
 
-TEST_F(PgSqlLeaseMgrTest, testSFLQ4) {
-    testSFLQ4();
+TEST_F(PgSqlLeaseMgrTest, testSflqCreateAndPick4) {
+    testSflqCreateAndPick4();
 }
 
-TEST_F(PgSqlLeaseMgrTest, testSFLQ6) {
-    testSFLQ6();
+TEST_F(PgSqlLeaseMgrTest, testSflqLeaseOps4) {
+    testSflqLeaseOps4();
+}
+
+TEST_F(PgSqlLeaseMgrTest, testSflqCreateAndPick6) {
+    testSflqCreateAndPick6();
+}
+
+TEST_F(PgSqlLeaseMgrTest, testSflqLeaseOps6) {
+    testSflqLeaseOps6();
 }
 
 /// @brief Test fixture class for testing @ref CfgDbAccessTest using PostgreSQL
